QuestionBanks permissions

+++ src/modules/QuestionBanks/containers/QuestionBanksList.js
@@ -7,7 +7,15 @@ import { useNavigate, useLocation } from "react-router-dom";
 import { useQuestionBanks } from "../hooks";
 import { LoadingSpinner } from "shared/components/LoadingSpinner";
 
-export const QuestionBanksList = () => {
+export const QuestionBanksList = (props) => {
+  const { hasPermission } = props;
+  const hasCreatePermission = hasPermission("QuestionBanks", "Create");
+  const hasEditPermission = hasPermission("QuestionBanks", "Edit");
+  const hasDeletePermission = hasPermission("QuestionBanks", "Delete");
+
+  const history = useNavigate();
+  const location = useLocation();
+
   const {
     questionBank,
     questionBanksQuery,
@@ -17,8 +25,7 @@ export const QuestionBanksList = () => {
     deleteQuestionBank,
   } = useQuestionBanks({ load: true });
   const { data, isLoading } = questionBanksQuery;
-  const history = useNavigate();
-  const location = useLocation();
+
   const onConfirm = () => {
     deleteQuestionBank.mutate(questionBank.questionBankId);
   };
@@ -30,12 +37,16 @@ export const QuestionBanksList = () => {
   const ActionButtons = ({ value }) => {
     return (
       <>
-        <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
-          <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
-        </Button>
-        <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
-          <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
-        </Button>
+        {
+          hasEditPermission && <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
+            <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
+          </Button>
+        }
+        {
+          hasDeletePermission && <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
+            <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
+          </Button>
+        }
       </>
     );
   };
@@ -45,27 +56,33 @@ export const QuestionBanksList = () => {
       {
         Header: "Name",
         accessor: "name",
-      },
-      {
-        Header: "Actions",
-        accessor: "questionBankId",
-        id: "actions",
-        Cell: ActionButtons,
-      },
+      }
     ],
     []
   );
 
+  if (hasCreatePermission || hasDeletePermission) {
+    columns.push({
+      Header: "Actions",
+      accessor: "questionBankId",
+      id: "actions",
+      Cell: ActionButtons,
+    })
+  }
+
   if (isLoading) {
     return <LoadingSpinner />;
   }
+
   return (
     <ContentLayout
       title={"Question Bank"}
       subtitle={"List"}
       breadcrumb={[{ label: "Question Bank" }]}
     >
-      <QuestionBanksFilter />
+      {
+        hasCreatePermission && <QuestionBanksFilter />
+      }
       <TableLayout columns={columns} data={data} />
       <ModalLayout
         isOpen={isConfirmDelete}
